# rag-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BKQuiz RAG Service",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/user-rag/ask": {
      "post": {
        "tags": [
          "user-rag"
        ],
        "summary": "Ask User Documents",
        "operationId": "ask_user_documents_api_v1_user_rag_ask_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-classroom-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Classroom-Id"
            }
          },
          {
            "name": "x-internal-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRagAskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRagAskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/user-rag/ask": {
      "post": {
        "tags": [
          "user-rag"
        ],
        "summary": "Ask User Documents",
        "operationId": "ask_user_documents_api_v2_user_rag_ask_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-classroom-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Classroom-Id"
            }
          },
          {
            "name": "x-internal-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRagAskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRagAskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ConversationMessage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserRagAskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 2,
            "title": "Question"
          },
          "topK": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Topk"
          },
          "documentIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentids"
          },
          "conversationHistory": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessage"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Conversationhistory"
          },
          "debug": {
            "type": "boolean",
            "title": "Debug",
            "default": false
          },
          "includeSystemDocuments": {
            "type": "boolean",
            "title": "Includesystemdocuments",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "UserRagAskRequest"
      },
      "UserRagAskResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Usage"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/UserRagSource"
            },
            "type": "array",
            "title": "Sources"
          },
          "insufficientContext": {
            "type": "boolean",
            "title": "Insufficientcontext",
            "default": false
          },
          "debug": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "model",
          "usage",
          "sources"
        ],
        "title": "UserRagAskResponse"
      },
...
      "UserRagSource": {
        "properties": {
          "sourceId": {
            "type": "string",
            "title": "Sourceid"
          },
          "documentId": {
            "type": "string",
            "title": "Documentid"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "sourceType": {
            "type": "string",
            "title": "Sourcetype"
          },
          "pageNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagenumber"
          },
          "slideNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slidenumber"
          },
          "chunkIndex": {
            "type": "integer",
            "title": "Chunkindex"
          },
          "heading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "textPreview": {
            "type": "string",
            "title": "Textpreview"
          }
        },
        "type": "object",
        "required": [
          "sourceId",
          "documentId",
          "filename",
          "sourceType",
          "pageNumber",
          "slideNumber",
          "chunkIndex",
          "heading",
          "score",
          "textPreview"
        ],
        "title": "UserRagSource"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}